pyproject.toml
[project]
name = "dio-desafio-api-bancaria-fastapi"
version = "0.1.0"
description = "API RESTful assíncrona para operações bancárias com FastAPI."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "pwdlib[argon2]>=0.2.1",
    "sqlalchemy>=2.0.36",
    "uvicorn[standard]>=0.32.0",
]

[dependency-groups]
dev = [
    "asgi-lifespan>=2.1.0",
    "httpx>=0.28.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=term-missing -q"
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_covered = false
